Create README.md
This repository is a detailed record of my learning journey in the field of Web Development. It is organized into various sections, each dedicated to a specific aspect of web development. The repository is structured to reflect my progression from understanding the basic building blocks of web development, such as HTML, CSS, and JavaScript, to mastering advanced libraries and frameworks like React and Node.js.

Sections
HTML, CSS, JavaScript: This section forms the foundation of the repository. It covers the fundamental building blocks of web development. It includes basic syntax, layout design, and interactivity. Each concept is demonstrated through simple code snippets and examples.

React: This section delves into the React library, a popular choice for building user interfaces. It includes detailed examples and code snippets demonstrating the use of components, state, props, hooks, and more. This section reflects my understanding of how to design and structure interactive web interfaces using React.

Node.js: This section is dedicated to my learning of Node.js, a runtime environment that allows JavaScript to be run on the server side. It includes my implementations of various server-side functionalities using Node.js.

Frontend Projects: This section showcases the projects I've completed using the above technologies. Each project is a practical application of the concepts I've learned and serves as a demonstration of my web development skills.

Learning Works: This section contains my practice works, exercises, and small projects that I've done while learning. It serves as a testament to my consistent practice and effort in mastering web development.

Future Plans
I am currently moving towards backend development and will be adding more content related to it in the future. This includes learning and implementing databases, server-side programming, API development, and more. This future content will reflect my progression into full-stack development.

Usage
This repository is open for anyone to explore. You can clone it to your local machine and navigate through the directories to delve into the topics you're interested in. Each file can be opened in your preferred IDE or text editor.